+++ tools/efixname/main.cpp
@@ -17,7 +17,7 @@ int main(int argc, char* argv[])
 	else
 	{
 		QTextCodec *ec = QTextCodec::codecForName("EUC-KR");
-		QString out = ec->toUnicode(argv[1], strlen(argv[1]));
+		QString out = ec->toUnicode(argv[1], (int) strlen(argv[1]));
 		QByteArray test = ec->fromUnicode(out);
 		if (test != argv[1])
 		{
